requirements.txt
aniso8601==8.0.0
click==7.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-MySQL==1.5.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
numpy==1.19.4
opencv-python==4.4.0.46
openpyxl==3.0.5
Pillow==8.0.1
PyMySQL==0.10.1
pypng==0.0.20
PyQRCode==1.2.1
pytz==2020.1
PyYAML==5.3.1
pyzbar==0.1.8
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
